web side of the job qa flow.

+++ views/job/view.php
@@ -53,7 +53,10 @@
 							<? if ($job->get('status') == 'available'): ?>
 								<a class="btn btn-mini" href="<?=$job->getUrl()?>/edit"><i class="icon-cog"></i> edit</a>
 							<? endif ?>
-							<? if ($job->get('status') != 'taken'): ?>
+							<? if ($job->get('status') == 'qa'): ?>
+								<a class="btn btn-mini" href="<?=$job->getUrl()?>/qa"><i class="icon-check"></i> finish</a>
+							<? endif ?>
+							<? if ($job->get('status') != 'taken' && $job->get('status') != 'qa'): ?>
 								<a class="btn btn-mini" href="<?=$job->getUrl()?>/delete"><i class="icon-remove"></i> delete</a>
 							<? endif ?>
 							<a class="btn btn-mini" href="/job/create/job:<?=$job->id?>"><i class="icon-repeat"></i> re-run</a>
